Orders/Cart/Categories refactored

+++ ComputerStore.DataAccess/CategoriesRepository.cs
@@ -1,48 +1,52 @@
 ﻿using ComputerStore.DataAccess.Entities;
-using ComputerStore.DataAccess;
 using ComputerStore.DataAccess.Interfaces;
 using Microsoft.EntityFrameworkCore;
-using System;
-using Npgsql.TypeMapping;
-using AutoMapper;
-using System.Runtime.Intrinsics.X86;
 
 namespace ComputerStore.DataAccess
 {
     public class CategoriesRepository : ICategoriesRepository
     {
-        private ApplicationDbContext _context;
+        private readonly ApplicationDbContext _context;
+        
         public CategoriesRepository(ApplicationDbContext context)
         {
             _context = context;
         }
+
         public async Task AddAsync(Category item)
         {
-            await _context.Categories.AddAsync(item);
+            if(item != null)
+                await _context.Categories.AddAsync(item);
+            else 
+                throw new ArgumentNullException(nameof(item)); 
         }
 
         public async Task DeleteAsync(string id)
         {
-            var item = await _context.Categories
-                .FirstOrDefaultAsync(c => c.Id == id);
-            if (item != null)
-                _context.Categories.Remove(item);
-            else throw new Exception("Invalid category item");
+            if (!string.IsNullOrEmpty(id))
+            {
+                var item = await _context.Categories
+                    .FirstOrDefaultAsync(c => c.Id == id);
+                if (item != null)
+                    _context.Categories.Remove(item);
+                else 
+                    throw new Exception("Invalid category item");
+            }
+            else throw new ArgumentNullException("category entity id");
         }
+
         public async Task UpdateAsync(Category item)
         {
             if(item != null && item.Id != null)
-            {
                 _context.Update(item);
-            }
-            else throw new Exception("Invalid category item");
+            else 
+                throw new ArgumentNullException("category entity");
         }
 
         public async Task<List<Category>> GetAsync(Func<Category, bool> predicate)
         {
             List<Category> result = _context.Categories.Where(predicate).ToList();
-            result ??= new List<Category>();
-            return result;
+            return result ?? new List<Category>();
         }
 
         public async Task<List<Category>> GetAsync()
@@ -59,7 +63,7 @@ namespace ComputerStore.DataAccess
             return item ?? throw new Exception("Category not found: id:" + id);
         }
 
-        public async Task<bool> IsExists(string id)
+        public async Task<bool> IsExistsAsync(string id)
         {
             bool isAny = false;
             if (!string.IsNullOrEmpty(id))
